==> analyses/06-IPW_ML/00-kmeans_DRC.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# IPTW-Prep: Spatial Cross Validation

## Partition DRC for Spatial Cross Validation

```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())
```
```{r}
#--------------------------------------------
# Purpose of this script is partition the DRC
# by geographic coordinates for spatial cross-validation
#--------------------------------------------

library(tidyverse)
library(sf)
set.seed(48)

DRCprov <- readRDS("data/map_bases/vivid_DRCprov.rds")
ge <- readRDS("data/raw_data/dhsdata/VivIDge.RDS")

coords <- ge %>% 
  dplyr::mutate(longnum = sf::st_coordinates(geometry)[,1],
                latnum = sf::st_coordinates(geometry)[,2])
sf::st_geometry(coords) <- NULL


#.............
# K-means Clustering of DRC
#.............
keda <- data.frame(k = seq(2, 200, by = 1))
keda$kmeans <- map(keda$k, function(k){return(kmeans(x = coords[,c("longnum", "latnum")], 
                                                     centers = k))})
keda$wss <- map(keda$kmeans, "withinss")
keda$totalwss <- map(keda$wss, function(x){return(sum(x))})


keda.df <- keda %>% 
  dplyr::select(c("k", "totalwss")) %>% 
  tidyr::unnest(cols = c("totalwss"))

kesteda.plotObj <- keda.df %>% 
  tibble::as_tibble(.) %>% 
  ggplot() +
  geom_line(aes(x=k, y=totalwss)) + 
  geom_point(aes(x=k, y=totalwss)) +
  geom_vline(xintercept = 15, color = "red", linetype = 2, alpha = 0.8) +
  theme_minimal() + 
  ylab("Total Within-Cluster Sum of Squares") +
  xlab("K")

jpeg("~/Documents/GitHub/VivID_Epi/results/figures/kmean_drc_paritition.jpg",
     width = 11, height = 8, units = "in", res = 500)
plot(kesteda.plotObj)
graphics.off()


```

```{r, results='asis'}
plot(kesteda.plotObj)
```
**K means of 15 appears to be the a good inflection point.**


```{r}
k <- kmeans(coords[,c("longnum", "latnum")], 15)
drcpart <- cbind.data.frame(coords, k = k$cluster) 

drcpart.plotObj <- drcpart %>% 
  ggplot() + 
  geom_sf(data = DRCprov) + 
  geom_point(aes(x=longnum, y=latnum, color = factor(k))) +
  theme(legend.position = "none")

jpeg("~/Documents/GitHub/VivID_Epi/results/figures/map_drc_paritition.jpg",
     width = 11, height = 8, units = "in", res = 500)
plot(drcpart.plotObj)
graphics.off()


```

```{r, results='asis'}
plot(drcpart.plotObj)
```


```{r}
drcpart <- drcpart %>% 
  dplyr::select(c("hv001", "k")) %>% 
  dplyr::rename(kmeansk = k)

saveRDS(drcpart, "data/derived_data/kmean_drcpartitioned.rds")

```

```{r}


jpeg("~/Documents/GitHub/VivID_Epi/results/figures/kmeans_map_drc_paritition.jpg",
     width = 11, height = 8, units = "in", res = 500)
cowplot::plot_grid(kesteda.plotObj, drcpart.plotObj, align = "h")
graphics.off()


```
