==> analyses/06-IPW_ML/01-IPTW_estimates.R <==
source("R/00-simple_Ensemble_Wrapper.R")
set.seed(48, "L'Ecuyer")
library(tidyverse)
library(mlr)
library(rslurm)

dt <- readRDS("data/derived_data/vividepi_recode.rds")
sf::st_geometry(dt) <- NULL
#........................
# manipulate tx map
#........................
# read in treatments
txs <- readRDS("model_datamaps/IPTW_treatments.RDS") %>% 
  dplyr::rename(target = column_name,
                positive = positivefactor) %>% 
  dplyr::mutate(type = ifelse(grepl("cont", target), "continuous",
                              ifelse(grepl("fctb", target), "binary", NA)))
  
#........................
# manipulate data
#........................
# subset to treatments, outcome, weights and coords
dt.ml <- dt %>% 
  dplyr::select(c("pv18s" , "pfldh", "hv005_wi", txs$target, 
                "longnum", "latnum"))

# subset to complete cases
dt.ml <- dt.ml %>% 
  dplyr::filter(complete.cases(.)) %>% 
  data.frame(.)


#........................
# Subset and Store Dataframes for Tasks
#........................
txs$data <- purrr::map2(.x = txs$target, .y = txs$adj_set, 
                        .f = function(x, y){
                          ret <- dt.ml %>% 
                            dplyr::select(c(x,y))
                          return(ret)
                        })



#-------------------------------------------------
# Run the Models
#-------------------------------------------------

# first make the tasks
txs$task <- purrr::pmap(txs[,c("data", "target", "positive", "type")], 
                         .f = make_class_task)

# now look and correct class imbalance
txs$task <- purrr::pmap(txs[,c("task", "type")], 
                         .f = find_Class_Imbalance,
                        classimb_tol = 0.6,
                        smotenn = 5)


# now make the ensemble learner
txs$learner <- purrr::map(txs$type, make_simple_Stack, 
                          learners = baselearners.list)

# make full model
# txs$fullmodel <- purrr::map2(.x = txs$learner, .y = txs$task,
#                             train)


# get predictions
# txs$preds <- purrr::map2(.x = txs$fullmodel, .y = txs$task,
#                             predict)

#.....................................
# SLURM 
#.....................................

slurm_trainpredict <- function(learner = learner, task=task){
  fullmodel <- mlr::train(learner = learner,
                          task = task)
  predictions <- predict(fullmodel, task = task)
  
  ret <- list(
    fullmodel = fullmodel,
    predictions = predictions
  )
  return(ret)
  
}

paramsdf <- txs[,c("learner", "task")]


ntry <- 18
sjob <- rlurm::slurm_apply(f = slurm_trainpredict, 
                    params = paramsdf, 
                    jobname = 'vivid_preds',
                    nodes = 18, 
                    cpus_per_node = 1, 
                    submit = F,
                    slurm_options = list(mem = 32000,
                                         array = sprintf("0-%d%%%d", 
                                                         ntry - 1, 
                                                         16),
                                         'cpus-per-task' = 8,
                                         error =  "%A_%a.err",
                                         output = "%A_%a.out",
                                         time = "5-00:00:00"))

# saveRDS(object = txs, file = "results/ensemble_predictive_probs.RDS")
