==> analyses/06-IPW_ML/02-check_weights.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Check Weights

```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())
```

```{r}
library(mlr)
library(tidyverse)
source("R/00-IPTW_functions.R")

```



### Distribution of IPTW for each Risk Factor

#### Table 
```{r checkwidist}

dcdr <- readxl::read_excel(path = "internal_datamap_files/risk_factor_plot_liftover.xlsx") # pretty decoder
#................................................................. 
# read in IPTW Estimates
#.................................................................
params <- readRDS("results/ensembl_cvRisk_paramdf.RDS")
params <- dplyr::left_join(params, dcdr, by = "target")
#......................
# look at weight
#......................
widist <- purrr::map(params$iptw, summary) %>% 
  do.call("rbind.data.frame", .) %>% 
  dplyr::mutate_if(is.numeric, round, 2) %>% 
  magrittr::set_colnames(c("min", "1stquart", "median", "mean", "3rdqart", "max"))

```

```{r, results='asis'}
cbind.data.frame(params$abridged_var_label, widist) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 30,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

### Boxplot
```{r checkwiboxplot}

# make plot
iptw.dist.plotObj <- params %>% 
  dplyr::select(c("abridged_var_label", "iptw")) %>% 
  tidyr::unnest(cols = iptw) %>% 
  dplyr::mutate(logiptw = log(iptw)) %>% 
  ggplot() + 
  geom_boxplot(aes(x = abridged_var_label, y = logiptw)) +
  geom_hline(yintercept = 0, col = "#e41a1c") +
  vivid_theme + 
  ylab("log(IPW)") +
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.text.x = element_text(family = "Helvetica", size = 10, angle = 90, hjust = 1, vjust = 0.5, face = "bold"),
        axis.line.x = element_line(size = 0.5, linetype = "solid", colour = "#000000"),
        axis.line.y = element_line(size = 0.5, linetype = "solid", colour = "#000000"))

jpeg("~/Documents/GitHub/VivID_Epi/results/figures/IPTW_wi_dist.jpg",
     width = 11, height = 8, units = "in", res = 500)
plot(iptw.dist.plotObj)
graphics.off()


```

```{r checkwiboxplotplot, results='asis'}

plot(iptw.dist.plotObj)

```

**N.B.** Although a view of our weights are more extreme than I would like, they are "as expected" for issues of smoothing through low likelihood exposure profiles. In other words, there are very few poor individuals with modern houses (as an example). Therefore, some of the extreme weight settings are valid. Moreover, our weights are 1 - 6 order of magnitudes less than those present in the [Gruber/Hernan 2015]() mansucript. In addition, hospital distance is extremely associated with urbanicity (as expected).


### Covariate Balance
A classic definition of confounding in a cross-sectional survey is an unequal distribution of baseline exposures. This will result in various covariates being correlated (e.g. being poor and living in a traditional house). If our weights have worked, we will have removed baseline confounding (i.e. removed baseline correlations). One way to check this is to measure the correlations of all the covariates at baseline (_Base_) as compared to covariate correlations with IPTW weights applied. Given that we have a mix of continuous and binary covariates, pearson correlation and/or standard-mean difference are not on the same squale. To overcome this, we will use the an *energy* function which calculates correlations despite functional form [Rizzo/Szekely 2016](https://onlinelibrary.wiley.com/doi/full/10.1002/wics.1375). This is more or less measuring correlation with euclidean distances of our data points in a multidimensional space (relating to kernel density and multi-dimensional GPs).     

Given that this fancy statistic doesn't take into account weights, I made 100 pseudopopulations (by sampling from the base population with respect to the calculated IPTW) and calculated the energy correlation for each iteration. Plotted is the distribution of the correlations for each covariate (e.g. Precipitation-Temperature, Precipitation-Wealth, etc. etc.) given that we focused on canonical sets when possible. 
```{r checkwicovarbalparams}
covarbalparams <- readRDS("analyses/06-IPW_ML/_rslurm_covar_IPTW_balance/params.RDS") 

covarretfiles <- list.files(path = "analyses/06-IPW_ML/_rslurm_covar_IPTW_balance/",
                            pattern = ".RDS", full.names = T)
covarretfiles <- covarretfiles[!grepl("f.RDS|params", covarretfiles)]

covarretfiles <- tibble::tibble(path = covarretfiles, 
                                ord = basename(covarretfiles)) %>% 
  dplyr::mutate(ord = stringr::str_extract(ord, "[0-9]+"),
                ord = as.numeric(ord)) %>% 
  dplyr::arrange(ord) 
covarretfiles <- covarretfiles$path

covarretfiles <- split(covarretfiles, 1:length(covarretfiles))

covarret <- lapply(covarretfiles, function(x){
  ret <- readRDS(x)
  ret <- tibble( var = names(ret), vals = unlist(ret))
  return(ret)
}) %>% 
  dplyr::bind_rows()

# add in distance observations
covarbalparams <- covarbalparams %>% 
  dplyr::mutate(dist = covarret$vals)


# make pretty plot
varnames <- params %>% 
  dplyr::select(c("target", "abridged_var_label")) %>% 
  dplyr::rename(covar1 = target)

covarbalanceplot <- covarbalparams %>% 
  dplyr::left_join(., varnames, by = "covar1") %>% 
  dplyr::mutate(lvl_fct = factor(lvl, levels = c("base", "IPTW"),
                                 labels = c("Base", "IPW"))) %>% 
  ggplot() + 
  geom_boxplot(aes(x = abridged_var_label, y = dist)) +
  facet_wrap(~lvl_fct, nrow = 2) +
  vivid_theme + 
  ylab("Correlation") +
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.text.x = element_text(family = "Helvetica", size = 10, angle = 90, hjust = 1, vjust = 0.5, face = "bold"),
        axis.line.x = element_line(size = 0.5, linetype = "solid", colour = "#000000"),
        axis.line.y = element_line(size = 0.5, linetype = "solid", colour = "#000000"))

jpeg("results/figures/IPTWcovar_bal.jpg",
     width = 11, height = 8, units = "in", res = 500)
plot(covarbalanceplot)
graphics.off()


```

```{r checkwicovarbalparamsplot, results='asis'}
plot(covarbalanceplot)
```

```{r, results='asis'}

covarbalparams_ranges <- covarbalparams %>% 
  dplyr::select(-c("data")) %>% 
  dplyr::group_by(covar1, lvl) %>% 
  dplyr::summarise(
    n = n(),
    maxcorr = max(dist),
    meancorr = mean(dist),
    mediancorr = median(dist),
    mincorr = min(dist)) 

DT::datatable(covarbalparams_ranges, extensions='Buttons',
              options = list(
                searching = T,
                pageLength = 20,
                dom = 'Bfrtip', 
                buttons = c('csv')))


```


**Very nice!** 

#### Covar Balance Long Comparison

Looking at descendent and ancestor not on causal path numbers. 
```{r, results='asis'}

covarbalparams %>% 
  dplyr::select(-c("data")) %>% 
  dplyr::filter(lvl == "base") %>% 
  dplyr::group_by(covar1) %>% 
  dplyr::summarise(
    n = n()) %>% 
  knitr::kable(.)


```

Looking at mean reductions (for the "fold" statement). 
```{r}

bal_sum_long <- covarbalparams %>% 
  dplyr::select(-c("data")) %>% 
  dplyr::group_by(covar1, lvl) %>% 
  dplyr::summarise(
    n = n(),
    mean_dist = mean(dist),
    se_dist = sd(dist)/n
  ) 


```
```{r, results='asis'}
DT::datatable(bal_sum_long, extensions='Buttons',
              options = list(
                searching = T,
                pageLength = 20,
                dom = 'Bfrtip', 
                buttons = c('csv')))
```

```{r}
bal_sum_wide <- bal_sum_long %>% 
  dplyr::select(-c("n", "se_dist")) %>% 
  tidyr::spread(., key = "lvl", value = "mean_dist") %>% 
  dplyr::mutate(base_vs_IPTW_reduction = base/IPTW)
```
```{r, results='asis'}
DT::datatable(bal_sum_wide, extensions='Buttons',
              options = list(
                searching = T,
                pageLength = 20,
                dom = 'Bfrtip', 
                buttons = c('csv')))

knitr::kable(cbind(nm = "base_vs_IPTW_reduction", summary(bal_sum_wide$base_vs_IPTW_reduction)))
```





### Appendix: Cross-Validated Risk Contributions
```{r checkwiappendix1}

#.......................................
# Pull out some relevant features
# from our fitted IPTW
#.......................................
stbmdldf <- params %>% 
  dplyr::select(c("target", "ensembl_cvRisk"))

stbmdldf$cvrisk.coef <- purrr::map(stbmdldf$ensembl_cvRisk, "cvrisk.coef")
stbmdldf$alg.cvrisk.validationset <- purrr::map(stbmdldf$ensembl_cvRisk, "alg.cvrisk.validationset")

stbmdldf$learners <- purrr::map(stbmdldf$alg.cvrisk.validationset, 
                                function(x){
                                  ret <- names(x)
                                  if(is.null(ret)){
                                    ret <- "regr"
                                  } 
                                  return(ret)
                                })

```

```{r checkwiappendix2, results='asis'}

out <- dplyr::left_join(dcdr, stbmdldf, by = "target") %>% 
  dplyr::select(-c("group")) %>% 
  dplyr::mutate_if(is.numeric, round, 2)

out %>% 
  dplyr::select(-c("ensembl_cvRisk")) %>% 
  tidyr::unnest(cols = c("cvrisk.coef", "alg.cvrisk.validationset", "learners")) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))


```
