==> analyses/06-IPW_ML/03-MSM_RRest.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# IPTW Risk Ratios

```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())

```

```{r}

library(srvyr)
options(survey.lonely.psu="adjust")
library(tidyverse)
library(ggdag)
source("R/00-functions_epi.R")
source("R/00-functions_basic.R")

```


## DAG
Note, the canonical set is all nodes in the DAG that are not on the casual pathway (i.e. "mediators"). By including the canonical set, we have taken a conservative approach for controlling for "lingering" or "unmeasured" confounders. 
  
```{r, results='asis', fig.align='center', fig.width=8, fig.height=6}
vividdag <- dagitty::downloadGraph(x = "dagitty.net/mjKNQhB")
tidy_dag <- ggdag::tidy_dagitty(vividdag)
# sub out vivax for malaria
tidy_dag$data$name[tidy_dag$data$name == "I:Pv18s"] <- "Malaria"

DAGplot <- ggdag::ggdag(tidy_dag) +
  ggdag::theme_dag()

dagplot <- tidy_dag %>% 
  ggplot(aes(x = x, y = y, xend = xend, yend = yend, color = name)) +
  geom_dag_point() +
  geom_dag_edges() +
  geom_dag_text(color = "#000000", size = 2, fontface = "bold") +
  theme_dag() +
  theme(legend.position = "none")



plot(dagplot)

jpeg(filename = "results/figures/dag.jpg",
     width = 11, height = 8, units = "in", res=500)
plot(dagplot)
graphics.off()


svglite::svglite("results/figures/dag.svg",
     width = 11, height = 8)
plot(dagplot)
graphics.off()

```

```{r}
# import data
dt <- readRDS("data/derived_data/vividepi_recode_completecases.rds")
dcdr <- readxl::read_excel(path = "internal_datamap_files/risk_factor_plot_liftover.xlsx") # pretty decoder
params <- readRDS("results/ensembl_cvRisk_paramdf.RDS")

iptw <- params %>% 
  dplyr::select(c("target", "iptw"))

```


```{r}

makeiptwsrvy <- function(x){ # not generalizable
  if(is.null(x)){ # if null, we didn't do weights was unconfounded in exp
    x <- 1
  }
  dt$wi <- x*dt$hiv05_wi
  dtsrvy <- dt %>% srvyr::as_survey_design(ids = hv001, 
                                           strata = hv023, 
                                           weights = wi)
  return(dtsrvy)
}


fitglmlogit <- function(outcome, target, iptwsrvy){
    iptwsrvy <- iptwsrvy %>% 
    dplyr::select(c(outcome, target))
  
  eq <- as.formula(paste0(outcome, "~", target))
  ret <- survey::svyglm(eq,
                        design = iptwsrvy,
                        family = quasibinomial(link="logit"))
  return(ret)
}

```


## _P. vivax_ Risk Factor IPTW-Prevelance Ratios
```{r}
#.......................
# Pvivax 
#.......................
pvivrskfctr <- dcdr$target
pvivrskfctr_models <- data.frame(outcome = rep("pv18s", length(pvivrskfctr)), 
                                 target = pvivrskfctr, stringsAsFactors=FALSE)

pvivrskfctr_models <- dplyr::left_join(pvivrskfctr_models, iptw, by = "target")


pvivrskfctr_models$iptwsrvy <- purrr::map(pvivrskfctr_models$iptw, makeiptwsrvy)


pvivrskfctr_models$glmlogit <- purrr::pmap(pvivrskfctr_models[,c("outcome", "target", "iptwsrvy")], fitglmlogit)

pvivrskfctr_models$glmlogit_tidy <- purrr::map(pvivrskfctr_models$glmlogit,
                                               .f=function(x){
                                                 broom::tidy(x, exponentiate=TRUE, conf.int=TRUE)}
)

pvivrskfctr_est <- pvivrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  mutate(species = "Pv")

```

```{r, results='asis'}
DT::datatable(pvivrskfctr_est, 
              rownames = F,
              options = list(
                searching = F,
                pageLength = 30,
                dom = 'Bfrtip', 
                buttons = c('csv')),
              caption = htmltools::tags$caption(
                style = 'caption-side: bottom; text-align: center;',
                "P vivax Risk Factor Table for IPTW Prevalence Odds Ratio"))


```

## _P. falciparum_ IPTW-Prevalence Odds Ratios Factor IPTW
```{r}
#.......................
# Pfalciparum 
#.......................
pfalrskfctr <- dcdr$target
pfalrskfctr_models <- data.frame(outcome = rep("pfldh", length(pfalrskfctr)), 
                                 target = pfalrskfctr, stringsAsFactors=FALSE)

pfalrskfctr_models <- dplyr::left_join(pfalrskfctr_models, iptw, by = "target")

pfalrskfctr_models$iptwsrvy <- purrr::map(pfalrskfctr_models$iptw, makeiptwsrvy)

pfalrskfctr_models$glmlogit <- purrr::pmap(pfalrskfctr_models[,c("outcome", "target", "iptwsrvy")], fitglmlogit)

pfalrskfctr_models$glmlogit_tidy <- purrr::map(pfalrskfctr_models$glmlogit,
                                             .f=function(x){
                                               broom::tidy(x, exponentiate=TRUE, conf.int=TRUE)}
)

pfalrskfctr_est <- pfalrskfctr_models$glmlogit_tidy %>% 
  bind_rows() %>% filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  mutate(species = "Pf")

```

```{r, results='asis'}
DT::datatable(pfalrskfctr_est, 
              rownames = F,
              options = list(
                searching = F,
                pageLength = 30,
                dom = 'Bfrtip', 
                buttons = c('csv')),
              caption = htmltools::tags$caption(
                style = 'caption-side: bottom; text-align: center;',
                "P falciparum Risk Factor Table for IPTW Prevalence Odds Ratio"))
```

## IPTW Prevalence Odds Ratio Plot
```{r}
#...................................................
# Tidy
#...................................................
pvivrskfctr_est <- pvivrskfctr_est %>% 
  bind_rows() %>% 
  filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::rename(target = term) %>% 
  dplyr::mutate(
    target = ifelse(grepl("_fctb_clst", target),
                    stringr::str_extract(target, "[ -~]+_fctb_clst"),
                    target),
    target = ifelse(!grepl("_fctb_clst", target) & grepl("_fctb", target),
                    stringr::str_extract(target, "[ -~]+_fctb"),
                    target),
    species = "Pv"
  ) %>% 
  dplyr::left_join(., dcdr, by = "target") 


# PF
pfalrskfctr_est <- pfalrskfctr_est %>% 
  bind_rows() %>% 
  filter(term != "(Intercept)") %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::rename(target = term) %>% 
  dplyr::mutate(
    target = ifelse(grepl("_fctb_clst", target),
                    stringr::str_extract(target, "[ -~]+_fctb_clst"),
                    target),
    target = ifelse(!grepl("_fctb_clst", target) & grepl("_fctb", target),
                    stringr::str_extract(target, "[ -~]+_fctb"),
                    target),
    species = "Pf"
  ) %>% 
  dplyr::left_join(., dcdr, by = "target") 



malriskest <- dplyr::bind_rows(pvivrskfctr_est, pfalrskfctr_est)



#...................................................
# Make Plots
#...................................................
liftover <- data.frame(label = 
                         c("Sex (Male)",
                           "Age",
                           "HIV (+)",
                           "Farmer",
                           "Wealth (Poor)",
                           "Education (Lower)",
                           "Num. House Members",
                           "Precipitation",
                           "Temperature",
                           "Altitude",
                           "Water Dist.",
                           "Housing Materials (Trad.)",
                           "ITN Use (No)",
                           "Urban",
                           "Hospital Dist. (Far)"),
                       labelnum = 1:15)

blocks <- data.frame(ymin = c(0, 0, 0),
                     ymax = c(3.5, 3.5, 3.5),
                     xmin = c(0.5, 7.5, 13.5), 
                     xmax = c(2.5, 11.5, 15.5)
)

text <- data.frame(x = c(1.5, 5.75, 9.5, 12.5, 14.5),
                   y = c(2.5, 2.5, 2.5, 2.5, 2.5),
                   label = 
                     c("Demographic",
                       "Individual",
                       "Ecological",
                       "Interventions",
                       "Urbanicity"))

malriskest <- malriskest %>% 
  dplyr::left_join(., liftover, by = "label")

finalMSMplot <- ggplot() +
  geom_rect(data = blocks, 
            aes(xmin = xmin, xmax = xmax,  ymin = ymin, ymax = ymax), 
            fill = "#d9d9d9",
            color = "#d9d9d9") +
  # por lines
  geom_hline(yintercept = 1, color = "#e41a1c", linetype = "dashed", size = 1) +
  geom_hline(yintercept = 0, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  geom_hline(yintercept = 2, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  geom_hline(yintercept = 3, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # label lines
  # geom_vline(xintercept = 1, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 2, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 3, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 4, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 5, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 6, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 7, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 8, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 9, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 10, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 11, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 12, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 13, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # geom_vline(xintercept = 14, color = "#bdbdbd", linetype = "dashed", size = 0.5) +
  # text
  geom_text(data = text,
            aes(x = x, y = y, label = label), 
            fontface = "bold",
            color = "#525252",
            size = 6) +
  geom_pointrange(data = malriskest,
                  aes(x = labelnum, y = estimate, 
                      ymin = conf.low, ymax = conf.high,
                      color = factor(species)), alpha = 0.8, size = 1.25,
                  position = position_dodge(width = 0.5)) + 
  scale_color_viridis_d("Species") +
  coord_flip() +
  scale_x_continuous(breaks = c(1:nrow(liftover)), 
                     labels = liftover$label) +
  scale_y_continuous(breaks = c(0, 1, 2, 3)) +
  ylab("Prevalence Odds Ratio") + 
  theme(
    plot.title =  element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 14),
    axis.title.x = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 16),
    axis.text.x = element_text(family = "Helvetica", color = "#000000", vjust = 0.5, hjust = 0.5, size = 14),
    axis.text.y = element_text(family = "Helvetica", color = "#000000", face = "bold", vjust = 0.5, hjust = 1, size = 14),
    axis.title.y = element_blank(),
    legend.title = element_text(family = "Helvetica", face = "bold", vjust = 0.5, hjust = 0.5, size = 16),
    legend.text = element_text(family = "Helvetica", vjust = 0.5, hjust = 0.5, size = 14, angle = 0),
    legend.position = "right",
    axis.line.x = element_line(color = "black", size = 1.5),
    axis.line.y = element_blank(),
    axis.ticks.y = element_blank(),
    panel.background = element_rect(fill = "transparent"),
    plot.background = element_rect(fill = "transparent"),
    panel.grid = element_blank(),
    panel.border = element_blank()
    )


jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/IPTWMSM_final_figure.jpg", width = 11, height = 8, units = "in", res = 800)
plot(finalMSMplot)
graphics.off()

jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/IPTWMSM_final_figure_long.jpg", width = 14, height = 8, units = "in", res = 800)
plot(finalMSMplot)
graphics.off()

```

```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'}
plot(finalMSMplot)
```

### Appendix: Bed Net Coding

##### LLIN Definition

```{r, results='asis'}

llinet <- readRDS("results/llin_hml20_sens_analysis_spensemble.rds")
broom::tidy(llinet, exponentiate=TRUE, conf.int=TRUE) %>% 
  DT::datatable(.)

```

### Appendix: Digging into _Pv_-Precipitation
```{r}
dtsrvyr.app <- makecd2013survey(survey = dt)

pvsrvyr <- dtsrvyr.app %>% 
  dplyr::mutate(count = 1) %>% 
  dplyr::group_by(hv001) %>% 
  srvyr::summarise(n = srvyr::survey_total(count, vartype = "se"), # same sampling weight to whole cluster
                   pv = srvyr::survey_mean(pv18s, vartype = "se"),
                   pf = srvyr::survey_mean(pfldh, vartype = "se"))


precip <- dt %>% 
  dplyr::group_by(hv001) %>% 
  dplyr::summarise(
    precip = mean(precip_mean_cont_clst) # clst has same precip)
  )

pvprecip <- dplyr::left_join(pvsrvyr, precip, by = "hv001") %>% 
  dplyr::select(-c(dplyr::ends_with("_se"))) %>% 
  dplyr::select(c("hv001", "n", "precip", dplyr::everything())) %>% 
  tidyr::gather(., key = "species", value = "prev", 4:5)

pvprecipplot <- pvprecip %>% 
  ggplot() +
  geom_point(aes(x=precip, y = prev, size = n)) +
  facet_wrap(~species) + 
  xlab("Precipitation (cm)") + 
  ylab("Prevalence") +
  theme_classic() + 
  theme(legend.position = "none",
        axis.text.x = element_text(angle = 90))

jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/pvprecip_figure.jpg", width = 11, height = 8, units = "in", res = 600)
plot(pvprecipplot)
graphics.off()

```

```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'}
plot(pvprecipplot)
```


### Appendix: Raster Digging into _Pv_-Precipitation 
```{r}
# load spatial coordinates
ge <- readRDS("data/raw_data/dhsdata/VivIDge.RDS") %>% 
  dplyr::mutate(longnum = sf::st_coordinates(geometry)[,1],
                latnum = sf::st_coordinates(geometry)[,2])
genosf <- ge
sf::st_geometry(genosf) <- NULL

# load precip raster
precipraster <- readRDS("data/derived_data/vividepi_precip_study_period_effsurface.rds") 

# bring in spatial and plot
pvprecipplot <- pvsrvyr %>% 
  dplyr::left_join(., genosf, by = "hv001")  %>% 
  ggplot() +
  ggspatial::layer_spatial(data = precipraster, aes(fill = stat(band1)),
                           alpha = 0.8) +
  geom_point(aes(x = longnum, y = latnum, color = pv, size = n),
             alpha = 0.6) +
  scale_fill_viridis_c("Precip.", na.value = NA) +
  scale_color_distiller("Pv Prev", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  theme_void() +
  theme(legend.position = "bottom")

jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/pvprecip_raster_figure.jpg", width = 11, height = 8, units = "in", res = 600)
plot(pvprecipplot)
graphics.off()

```

```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'}
plot(pvprecipplot)
```




<!-- ### Appendix: Digging into _Pv_-Temperature -->
<!-- ```{r} -->

<!-- dtsrvyr.app <- makecd2013survey(survey = dt) -->

<!-- pvsrvyr <- dtsrvyr.app %>%  -->
<!--   dplyr::group_by(hv001) %>%  -->
<!--   dplyr::mutate(count = 1) %>%  -->
<!--   srvyr::summarise(n = srvyr::survey_total(count),  -->
<!--                    pv = srvyr::survey_mean(pv18s, vartype = "se"), -->
<!--                    pf = srvyr::survey_mean(pfldh, vartype = "se")) -->

<!-- temp <- dt %>%  -->
<!--   dplyr::group_by(hv001) %>%  -->
<!--   dplyr::summarise( -->
<!--     temp = mean(temp_mean_cont_clst) # clst has same precip) -->
<!--   ) -->

<!-- tempprecip <- dplyr::left_join(pvsrvyr, temp, by = "hv001") %>%  -->
<!--   dplyr::select(-c(dplyr::ends_with("_se"))) %>%  -->
<!--   dplyr::select(c("hv001", "n", "temp", dplyr::everything())) %>%  -->
<!--   tidyr::gather(., key = "species", value = "prev", 4:5) -->

<!-- pvtempplot <- tempprecip %>%  -->
<!--   ggplot() + -->
<!--   geom_point(aes(x=temp, y = prev, size = n)) + -->
<!--   facet_wrap(~species) +  -->
<!--   xlab("Temperature (C)") +  -->
<!--   ylab("Prevalence") + -->
<!--   theme_classic() +  -->
<!--   theme(legend.position = "none", -->
<!--         axis.text.x = element_text(angle = 90)) -->

<!-- jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/pvtemp_figure.jpg", width = 11, height = 8, units = "in", res = 600) -->
<!-- plot(pvtempplot) -->
<!-- graphics.off() -->

<!-- ``` -->

<!-- ```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'} -->
<!-- plot(pvtempplot) -->
<!-- ``` -->


<!-- ### Appendix: Raster Digging into _Pv_-Temperature  -->
<!-- ```{r} -->
<!-- # load spatial coordinates -->
<!-- ge <- readRDS("data/raw_data/dhsdata/VivIDge.RDS") -->

<!-- # load temp raster -->
<!-- tempraster <- readRDS("data/derived_data/vividepi_temperature_study_period_effsurface.rds")  -->

<!-- # bring in spatial and plot -->
<!-- pvtempplot <- pvsrvyr %>%  -->
<!--   dplyr::left_join(., genosf, by = "hv001") %>%  -->
<!--   ggplot() + -->
<!--   ggspatial::layer_spatial(data = tempraster, aes(fill = stat(band1)), -->
<!--                            alpha = 0.8) + -->
<!--   geom_point(aes(x = longnum, y = latnum, color = pv, size = n), -->
<!--              alpha = 0.6) + -->
<!--   scale_fill_viridis_c("Temp", na.value = NA) + -->
<!--   scale_color_distiller("Pv Prev", type = "div", palette = "RdYlBu", -->
<!--                        na.value = NA) + -->
<!--   theme_void() + -->
<!--   theme(legend.position = "bottom") -->

<!-- jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/pvtemp_raster_figure.jpg", width = 11, height = 8, units = "in", res = 600) -->
<!-- plot(pvtempplot) -->
<!-- graphics.off() -->

<!-- ``` -->

<!-- ```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'} -->
<!-- plot(pvtempplot) -->
<!-- ``` -->





### Appendix: Digging into _Pv_-Health Distance
```{r, results='asis'}

dtsrvyr.app %>% 
  dplyr::group_by(hlthdist_fctb_clst, pv18s_fctb) %>% 
  dplyr::summarise(counts = srvyr::survey_total()) %>% 
  knitr::kable(.)

```


### Appendix: Raster Digging into _Pv_-Health Distance
```{r}
# load spatial coordinates
ge <- readRDS("data/raw_data/dhsdata/VivIDge.RDS")
# create bounding box of Central Africa for Speed
caf <- as(raster::extent(10, 40,-18, 8), "SpatialPolygons")
sp::proj4string(caf) <- "+init=epsg:4326"

# create mask 
DRC <- readRDS("data/map_bases/gadm/gadm36_COD_0_sp.rds")

# load hlthdist raster
hlthdistraster <- raster::raster("data/raw_data/hlthdist/2020_walking_only_travel_time_to_healthcare.geotiff")
sp::identicalCRS(hlthdistraster, caf)
sp::identicalCRS(hlthdistraster, sf::as_Spatial(DRC))
hlthdistraster <- raster::crop(hlthdistraster, caf)
hlthdistraster <- raster::mask(hlthdistraster, DRC)
hlthdistraster <- raster::aggregate(hlthdistraster, fact = 10) # agg for easier plotting 

# bring in spatial and plot
pvhlthdistplot <- pvsrvyr %>% 
  dplyr::left_join(., genosf, by = "hv001") %>% 
  ggplot() +
  ggspatial::layer_spatial(data = hlthdistraster, aes(fill = stat(band1)),
                           alpha = 0.8) +
  geom_point(aes(x = longnum, y = latnum, color = pv, size = n),
             alpha = 0.6) +
  scale_fill_viridis_c("Health Dist.", na.value = NA) +
  scale_color_distiller("Pv Prev", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  theme_void() +
  theme(legend.position = "bottom")

jpeg(filename = "~/Documents/GitHub/VivID_Epi/results/figures/pvhlthdist_raster_figure.jpg", width = 11, height = 8, units = "in", res = 600)
plot(pvhlthdistplot)
graphics.off()

```

```{r, results='asis', fig.width=11, fig.height=8, fig.align='center'}
plot(pvhlthdistplot)
```




### Appendix: Unadjusted vs. Adjusted RR Estimates
```{r}
load("~/Documents/GitHub/VivID_Epi/results/bivariate_model_results.rda")

pvconf <- pvivrskfctr_models$glmlogit_tidy %>% 
  dplyr::bind_rows() %>% 
  dplyr::filter(term != "(Intercept)") %>% 
  dplyr::select(c("term", "estimate", "conf.low", "conf.high")) %>% 
  dplyr::mutate(species = "Pv") %>% 
  dplyr::rename(target = term) %>% 
  dplyr::mutate(
     target = ifelse(grepl("_fctb_clst", target),
                         stringr::str_extract(target, "[ -~]+_fctb_clst"),
                         target),
    target = ifelse(!grepl("_fctb_clst", target) & grepl("_fctb", target),
                         stringr::str_extract(target, "[ -~]+_fctb"),
                         target)
  ) %>% 
  mutate_if(is.numeric, round, 2)


pfconf <- pfalrskfctr_models$glmlogit_tidy %>% 
  dplyr::bind_rows() %>% 
  dplyr::filter(term != "(Intercept)") %>% 
  dplyr::select(c("term", "estimate", "conf.low", "conf.high")) %>% 
  dplyr::mutate(species = "Pf") %>% 
  dplyr::rename(target = term) %>% 
  dplyr::mutate(
     target = ifelse(grepl("_fctb_clst", target),
                         stringr::str_extract(target, "[ -~]+_fctb_clst"),
                         target),
    target = ifelse(!grepl("_fctb_clst", target) & grepl("_fctb", target),
                         stringr::str_extract(target, "[ -~]+_fctb"),
                         target)
  ) %>% 
  mutate_if(is.numeric, round, 2)


confmalriskest <- rbind.data.frame(pvconf, pfconf) %>% 
  magrittr::set_colnames(c("target", "est_conf", "conf.low_conf", "conf.high_conf", "species"))

iptwest <- malriskest %>% 
  dplyr::select(c("target", "estimate", "conf.low", "conf.high", "species")) %>% 
  magrittr::set_colnames(c("target", "est_iptw", "conf.low_iptw", "conf.high_iptw", "species"))



out <- left_join(iptwest, confmalriskest, by = c("target", "species")) %>% 
  dplyr::select(c("species", dplyr::everything()))




out <- dplyr::left_join(dcdr, out, by = "target") %>% 
  dplyr::select(-c("group")) %>% 
  dplyr::arrange(dplyr::desc(species))
```

```{r, results='asis'}
DT::datatable(out, 
              rownames = F,
              options = list(
                searching = F,
                pageLength = 30,
                dom = 'Bfrtip', 
                buttons = c('csv'))
            #  , caption = htmltools::tags$caption(
            #    style = 'caption-side: bottom; text-align: center;',
            #    "Malaria Risk Factor Table with Unadjusted (Bivariate Assoc.) and IPTW Prevalence Odds Ratio")
              )
```
